pyproject.toml
[project]
name = "panodreamer"
version = "0.1.0"
description = "PanoDreamer: Optimization-Based Single Image to 360 3D Scene With Diffusion"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "transformers",
    "diffusers",
    "torch",
    "torchvision",
    "pillow",
    "tqdm",
    "accelerate",
    "kornia",
    "opencv-python",
    "matplotlib",
    "scikit-learn",
    "ropwr",
    "gsplat",
    "gdown",
    "imageio",
    "imageio-ffmpeg",
    "plyfile",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
